mbuswebd/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
attrs==24.2.0
bidict==0.23.1
blinker==1.8.2
click==8.1.7
dominate==2.9.1
Flask==3.0.3
Flask-Bootstrap==3.3.7.1
flask-nav==0.6
frozenlist==1.4.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
Hypercorn==0.17.3
hyperframe==6.0.1
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
multidict==6.0.5
nats-py==2.9.0
priority==2.0.0
python-engineio==4.9.1
python-socketio==5.11.3
PyYAML==6.0.2
Quart==0.19.6
quart-cors==0.7.0
quart-flask-patch==0.3.0
simple-websocket==1.0.0
uvicorn==0.30.6
visitor==0.1.3
Werkzeug==3.0.4
wsproto==1.2.0
yarl==1.9.4
